move function to load config from file
